requirements.txt
discord.py
PyNaCl
python-dotenv

gTTS
